requirements.txt
hx711==1.1.2.3
RPi.GPIO==0.7.1
rpimotorlib==3.2
